+++ app/policies/project_policy.rb
@@ -89,7 +89,7 @@ class ProjectPolicy
   end
 
   private def project_assigned_to_user?
-    @record.assigned_to == @user
+    @record.assigned_to == @user || @user.is_service_support?
   end
 
   private def edit_project_closed?